+++ pyrsa/rdm/__init__.py
@@ -1,11 +1,11 @@
 from .rdms import RDMs
 from .rdms import concat
 from .rdms import get_categorical_rdm
-from .rdms import rank_transform
-from .rdms import sqrt_transform
-from .rdms import positive_transform
 from .rdms import load_rdm
 from .rdms import rdms_from_dict
+from .transform import rank_transform
+from .transform import sqrt_transform
+from .transform import positive_transform
 from .calc import calc_rdm
 from .calc import calc_rdm_movie
 from .calc import calc_rdm_euclid
